requirements.txt
wheel
nbresult
colorama
ipdb
ipython
ipykernel
jupyterlab
yapf
matplotlib
pygeohash
pytest
pylint
seaborn
numpy==1.23.5
pandas==1.5.3
scipy==1.10.0
scikit-learn==1.2.1
google-cloud-bigquery
google-cloud-storage

db-dtypes
pyarrow

python-dotenv
psycopg2-binary

tensorflow-macos; sys_platform == 'darwin' and 'ARM' in platform_version # Mac M-chips
tensorflow; sys_platform == 'darwin' and 'ARM' not in platform_version # Mac Intel chips
tensorflow; sys_platform != 'darwin' # Windows & Ubuntu Intel chips

fastapi
pytz
uvicorn

